+++ packages/ast-spec/src/declaration/ImportDeclaration/spec.ts
@@ -1,12 +1,12 @@
 import type { AST_NODE_TYPES } from '../../ast-node-types';
 import type { BaseNode } from '../../base/BaseNode';
+import type { StringLiteral } from '../../expression/literal/StringLiteral/spec';
 import type { ImportClause } from '../../unions/ImportClause';
-import type { Literal } from '../../unions/Literal';
 import type { ImportKind } from '../ExportAndImportKind';
 
 export interface ImportDeclaration extends BaseNode {
   type: AST_NODE_TYPES.ImportDeclaration;
-  source: Literal;
+  source: StringLiteral;
   specifiers: ImportClause[];
   importKind: ImportKind;
 }