+++ packages/eslint-plugin/src/rules/class-literal-property-style.ts
@@ -44,7 +44,6 @@ export default util.createRule<Options, MessageIds>({
     docs: {
       description:
         'Ensures that literals on classes are exposed in a consistent style',
-      category: 'Best Practices',
       recommended: false,
     },
     fixable: 'code',
@@ -100,7 +99,7 @@ export default util.createRule<Options, MessageIds>({
     }
 
     return {
-      ClassProperty(node: TSESTree.ClassProperty): void {
+      PropertyDefinition(node: TSESTree.PropertyDefinition): void {
         if (!node.readonly || node.declare) {
           return;
         }