+++ packages/eslint-plugin/src/rules/indent.ts
@@ -9,15 +9,17 @@ import {
   TSESTree,
   AST_NODE_TYPES,
 } from '@typescript-eslint/experimental-utils';
-import baseRule from 'eslint/lib/rules/indent';
+import { getESLintCoreRule } from '../util/getESLintCoreRule';
 import * as util from '../util';
+
+const baseRule = getESLintCoreRule('indent');
 
 type Options = util.InferOptionsTypeFromRule<typeof baseRule>;
 type MessageIds = util.InferMessageIdsTypeFromRule<typeof baseRule>;
 
 const KNOWN_NODES = new Set([
   // Class properties aren't yet supported by eslint...
-  AST_NODE_TYPES.ClassProperty,
+  AST_NODE_TYPES.PropertyDefinition,
 
   // ts keywords
   AST_NODE_TYPES.TSAbstractKeyword,
@@ -33,7 +35,7 @@ const KNOWN_NODES = new Set([
   AST_NODE_TYPES.TSNullKeyword,
 
   // ts specific nodes we want to support
-  AST_NODE_TYPES.TSAbstractClassProperty,
+  AST_NODE_TYPES.TSAbstractPropertyDefinition,
   AST_NODE_TYPES.TSAbstractMethodDefinition,
   AST_NODE_TYPES.TSArrayType,
   AST_NODE_TYPES.TSAsExpression,
@@ -65,7 +67,6 @@ const KNOWN_NODES = new Set([
   AST_NODE_TYPES.TSModuleDeclaration,
   AST_NODE_TYPES.TSNonNullExpression,
   AST_NODE_TYPES.TSParameterProperty,
-  AST_NODE_TYPES.TSParenthesizedType,
   'TSPlusToken',
   AST_NODE_TYPES.TSPropertySignature,
   AST_NODE_TYPES.TSQualifiedName,
@@ -90,17 +91,14 @@ export default util.createRule<Options, MessageIds>({
     type: 'layout',
     docs: {
       description: 'Enforce consistent indentation',
-      category: 'Stylistic Issues',
       // too opinionated to be recommended
       recommended: false,
       extendsBaseRule: true,
     },
     fixable: 'whitespace',
+    hasSuggestions: baseRule.meta.hasSuggestions,
     schema: baseRule.meta.schema,
-    messages: baseRule.meta.messages ?? {
-      wrongIndentation:
-        'Expected indentation of {{expected}} but found {{actual}}.',
-    },
+    messages: baseRule.meta.messages,
   },
   defaultOptions: [
     // typescript docs and playground use 4 space indent
@@ -138,7 +136,7 @@ export default util.createRule<Options, MessageIds>({
         | TSESTree.TSEnumMember
         | TSESTree.TypeElement,
       type:
-        | AST_NODE_TYPES.ClassProperty
+        | AST_NODE_TYPES.PropertyDefinition
         | AST_NODE_TYPES.Property = AST_NODE_TYPES.Property,
     ): TSESTree.Node | null {
       const base = {
@@ -170,7 +168,7 @@ export default util.createRule<Options, MessageIds>({
           readonly: false,
           declare: false,
           ...base,
-        } as TSESTree.ClassProperty;
+        } as TSESTree.PropertyDefinition;
       }
     }
 
@@ -332,8 +330,8 @@ export default util.createRule<Options, MessageIds>({
             p =>
               TSPropertySignatureToProperty(
                 p,
-                AST_NODE_TYPES.ClassProperty,
-              ) as TSESTree.ClassProperty,
+                AST_NODE_TYPES.PropertyDefinition,
+              ) as TSESTree.PropertyDefinition,
           ),
 
           // location data