+++ packages/eslint-plugin/src/rules/no-confusing-void-expression.ts
@@ -30,7 +30,6 @@ export default util.createRule<Options, MessageId>({
     docs: {
       description:
         'Requires expressions of type void to appear in statement position',
-      category: 'Best Practices',
       recommended: false,
       requiresTypeChecking: true,
     },
@@ -71,6 +70,7 @@ export default util.createRule<Options, MessageId>({
     ],
     type: 'problem',
     fixable: 'code',
+    hasSuggestions: true,
   },
   defaultOptions: [{}],