+++ packages/eslint-plugin/src/rules/no-floating-promises.ts
@@ -22,11 +22,11 @@ export default util.createRule<Options, MessageId>({
   meta: {
     docs: {
       description: 'Requires Promise-like values to be handled appropriately',
-      category: 'Best Practices',
       recommended: 'error',
       suggestion: true,
       requiresTypeChecking: true,
     },
+    hasSuggestions: true,
     messages: {
       floating: 'Promises must be handled appropriately.',
       floatingVoid: