+++ packages/eslint-plugin/src/rules/no-inferrable-types.ts
@@ -20,7 +20,6 @@ export default util.createRule<Options, MessageIds>({
     docs: {
       description:
         'Disallows explicit type declarations for variables or parameters initialized to a number, string, or boolean',
-      category: 'Best Practices',
       recommended: 'error',
     },
     fixable: 'code',
@@ -195,7 +194,7 @@ export default util.createRule<Options, MessageIds>({
       node:
         | TSESTree.VariableDeclarator
         | TSESTree.Parameter
-        | TSESTree.ClassProperty,
+        | TSESTree.PropertyDefinition,
       typeNode: TSESTree.TSTypeAnnotation | undefined,
       initNode: TSESTree.Expression | null | undefined,
     ): void {
@@ -253,7 +252,9 @@ export default util.createRule<Options, MessageIds>({
       });
     }
 
-    function inferrablePropertyVisitor(node: TSESTree.ClassProperty): void {
+    function inferrablePropertyVisitor(
+      node: TSESTree.PropertyDefinition,
+    ): void {
       // We ignore `readonly` because of Microsoft/TypeScript#14416
       // Essentially a readonly property without a type
       // will result in its value being the type, leading to
@@ -269,7 +270,7 @@ export default util.createRule<Options, MessageIds>({
       FunctionExpression: inferrableParameterVisitor,
       FunctionDeclaration: inferrableParameterVisitor,
       ArrowFunctionExpression: inferrableParameterVisitor,
-      ClassProperty: inferrablePropertyVisitor,
+      PropertyDefinition: inferrablePropertyVisitor,
     };
   },
 });