+++ packages/eslint-plugin/src/rules/no-invalid-this.ts
@@ -2,12 +2,14 @@ import {
   TSESTree,
   AST_NODE_TYPES,
 } from '@typescript-eslint/experimental-utils';
-import baseRule from 'eslint/lib/rules/no-invalid-this';
+import { getESLintCoreRule } from '../util/getESLintCoreRule';
 import {
   InferOptionsTypeFromRule,
   createRule,
   InferMessageIdsTypeFromRule,
 } from '../util';
+
+const baseRule = getESLintCoreRule('no-invalid-this');
 
 export type Options = InferOptionsTypeFromRule<typeof baseRule>;
 export type MessageIds = InferMessageIdsTypeFromRule<typeof baseRule>;
@@ -19,13 +21,14 @@ export default createRule<Options, MessageIds>({
     docs: {
       description:
         'Disallow `this` keywords outside of classes or class-like objects',
-      category: 'Best Practices',
       recommended: false,
       extendsBaseRule: true,
     },
+    // TODO: this rule has only had messages since v7.0 - remove this when we remove support for v6
     messages: baseRule.meta.messages ?? {
       unexpectedThis: "Unexpected 'this'.",
     },
+    hasSuggestions: baseRule.meta.hasSuggestions,
     schema: baseRule.meta.schema,
   },
   defaultOptions: [{ capIsConstructor: true }],
@@ -50,10 +53,10 @@ export default createRule<Options, MessageIds>({
 
     return {
       ...rules,
-      ClassProperty(): void {
+      PropertyDefinition(): void {
         thisIsValidStack.push(true);
       },
-      'ClassProperty:exit'(): void {
+      'PropertyDefinition:exit'(): void {
         thisIsValidStack.pop();
       },
       FunctionDeclaration(node: TSESTree.FunctionDeclaration): void {