+++ packages/shared-fixtures/fixtures/typescript/basics/abstract-class-with-abstract-properties.src.ts
@@ -1,4 +1,4 @@
 abstract class Foo {
     abstract bar;
-    abstract baz = 3;
+    abstract baz: number;
 }