+++ packages/types/tools/copy-ast-spec.ts
@@ -11,7 +11,7 @@ const AST_SPEC_PATH = path.resolve(__dirname, '../../ast-spec');
 const OUTPUT_PATH = path.join(path.resolve(__dirname, '../src/'));
 
 // ensure the package is built
-chlidProcess.execSync('yarn build', { cwd: AST_SPEC_PATH });
+chlidProcess.execSync('yarn build', { cwd: AST_SPEC_PATH, stdio: 'inherit' });
 
 const HEADER = `\
 /**********************************************