+++ packages/typescript-estree/tests/ast-alignment/spec.ts
@@ -1,7 +1,12 @@
 import fs from 'fs';
+import type { File } from '@babel/types';
 import { fixturesToTest } from './fixtures-to-test';
 import { parse } from './parse';
-import * as parseUtils from './utils';
+import {
+  preprocessBabylonAST,
+  preprocessTypescriptAST,
+  removeLocationDataAndSourceTypeFromProgramNode,
+} from './utils';
 
 fixturesToTest.forEach(fixture => {
   const filename = fixture.filename;
@@ -72,13 +77,13 @@ fixturesToTest.forEach(fixture => {
      * Perform some extra formatting steps on the babel AST before comparing
      */
     expect(
-      parseUtils.removeLocationDataAndSourceTypeFromProgramNode(
-        parseUtils.preprocessBabylonAST(babelParserResult.ast),
+      removeLocationDataAndSourceTypeFromProgramNode(
+        preprocessBabylonAST(babelParserResult.ast as File),
         fixture.ignoreSourceType,
       ),
     ).toEqual(
-      parseUtils.removeLocationDataAndSourceTypeFromProgramNode(
-        parseUtils.preprocessTypescriptAST(typeScriptESTreeResult.ast),
+      removeLocationDataAndSourceTypeFromProgramNode(
+        preprocessTypescriptAST(typeScriptESTreeResult.ast),
         fixture.ignoreSourceType,
       ),
     );